ARM: add tegra support to oprofile, fix cpu_mask conventions following merge
the ARM-2.6.28 tree's version of oprofile requires that the base address and
IRQ for the performance monitor, l2 cache controller and SCU be defined for
the platform or compilation fails. since this is platform-specific, a new
set of defines is added for Tegra.

additionally, when the arm-2.6.28 stable branch was merged on top of
the 2.6.29 kernel, a compilation failure resulted due to irq_set_affinity
expecting a cpu_mask* and oprofile providing cpu_mask.

bug: 637617
Change-Id: I13e4851fb7a99d0dbb1becfbf62ad79fb9fe48a0