tegra power: Add initial LP0 implementation.
The function enter_lp0 puts SDRAM into self-refresh, and puts the system
into the LP0 state via the SIDE_EFFECT bit. This is needed to initiate
a warmboot upon return the deep sleep state.

Change the exit_power_state function to account for the possibility
of returning via LP0 instead of LP2. exit_power_state is the common
return point for LP2 and LP0. It must be able to distinguish
between the two states.

Change-Id: Ifb93a4346380f44322517990ed71acab56e1fc8e