tegra: Handling the arbitration lost error in i2c.
When I2c master start the address cycle, it tries to get the control of
the bus but if bus is taken by some other master or slave have not released
the SDA line, the i2c controller reports the arbitration lost error and
transaction aborts by the i2c controller. The i2c controller generates
the interrupt if it is enabled. Driver handles this error by clearing the
status and return to the caller with ArbitrationFailed.

Flag is added to enable the polling/interrupt method of the transaction.

Bug 634931 I2C handling in i2cisr.
Bug 622224 Ap20 whistler: Sometimes pmu stop responding when the WIFI card is inserted on board.
Tested on Harmony.

Change-Id: Ic5260c24e0888beb974a7c55c05d2ef7b939f3c3