tegra: allow native serial driver uarts to be registered
init_common was using a preprocessor conditional to compile-out the
uart platform device registration code, but was only checking
for the ddk-based uart driver (CONFIG_SERIAL_TEGRA_DDK), not the
native one (CONFIG_SERIAL_TEGRA).

Furthermore, the registration was registering all controllers
on the chip, not the number of valid UARTs on the system.

Change-Id: I49b96412c27deed1afe039cfecf2bede681cf320