serial: compile fixes for tegra native serial driver
when the serial driver was merged to git from perforce, some errors
were introduced during the reformatting of it. these are now fixed.

add a Kconfig to specify the maximum number of high-speed UARTs
supported by the kernel, which will be used to ensure that enough
DMA channels are reserved from the Tegra RM to allow the driver
initialization to succeed

Change-Id: I363a8375ebbdda004161672ff3700eb09be474ae