tegra RM: Fix PWM scaling calculation in RM for backlight intensity.
Fixed PWM scaling calculation and added boundary check. This is an 8-bit
value that was being scaled to a max of 256 instead of 255, causing an
overflow in the boundary case of a maximum setting (thereby resulting in 0).

Bug 638387 [Harmony] Backlight is broken and there is no display.
Bug 627575 [firefly2/Harmony/Display]External display panel not adjusting
to Brightness change from display settings.
Tested on: Android, Harmony