tegra2 rtc: date/time non-persistent on boards
Multi-stage Changes are as follows:
Change1:
1) TPS6586x PMU RTC(harmony) cannot store more than 34 years duration.
 Hence, RTC driver has been modified to use year 2009 as reference
 instead of the year 1970(used in linux by default). Change in
 tps6586x rtc implementation.
 Issue was Fixed in Bug 621031 earlier and checked into p4 by modifying
 tegra rtc driver. We are using a different approach now by only
 modifying tps6586x rtc driver. This way rtc drivers for other PMU
 like max8907b which can store as much 9999 years in RTC do not need
 to change.
2) MAX8907B PMU RTC(whistler) driver was not storing day/month/year data
 earlier. Further, on every boot the RTC was being cleared. Fixed this
 in file max8907b_rtc.c
 Implementation specific to linux as uses library functions: mktime and
 rtc_time_to_tm
3) MAX8907B PMU I2C read API implementation was not reading 4th byte.
 Changed this as DD, MM, YY1, YY2 information is being read now.
 Changed max8907b_i2c.c Further, changes are as follows:
Change2:
1) TPS6586x PMU RTC(harmony) driver modified to select crystal oscillator
as source for RTC. After this change find that RTC counts even when
harmony is powered off.
2) MAX8907B PMU RTC(whistler) driver modified to incorporate review
comments
a) renamed macro SHIFT_TO_2009 as OFFSET_BASE_YEAR
b) used a macro for BCD to decimal conversion for better readability.

Change set 1 was committed but was reverted because of bugs 645312
and 645286. Further tests showed that if the date is set before
testing gallery app the issue is not seen. Trying to re-submit
reverted checkin for change1 along with change2 as a new change.

Change3:
a) decimal to bcd expresssion converted to macro
b) corrected spelling

Bug 625990 : [Firefly2/Harmony] Date & Time is not persisting after reboot

Tested on : Harmony + whistler. Date/time persists.
	Note: Whistler rechargable battery needs to be charged to 2.4V for
	RTC data to persist. Harmony has non-rechargable battery.

Change-Id: Icc543261096ee4c89cc1e394e2b3868ac370f8e9