tegra RM: add nvmem= command-line parsing for carveout aperture(s)
the bootloader may place carveout in a location that is not exactly at
the top of physical memory (if, for example, the bootloader does not
support mapping all the way to the top of memory), and the existing code
in the RM initialization did not allow for this (or for multiple carveout
apertures).

add a new command-line parameter (nvmem=) to position the carveout aperture
arbitrarily in the physical address space.

Change-Id: Ib750c4c038916a21b9fece490efbe6c953da09de