tegra nvos: return interrupted semaphores to user-land
split the nvos interrupt implementation into 2 parts: one which returns
an error code when the thread is awakened due to a pending signal, and
one which loops in the kernel until the semaphore operation completes.

the user-land stub needs to immediately return control to the application
in the event of a pending signal, to ensure that the signal is dealt
with before re-attempting to wait on the semaphore.

bug: 642544
Change-Id: I0b233c9483d67cc6d25285d0dc3eddafa8502500