tegra: Update the UART Tx data path and added new system DMA API.
-Using workqueues instead of tasklets for Tx DMA.
-Waiting for the FIFO to drain after the DMA is done, as it is apparently
needed by the HW.
- Added new DMA API to check if the req is already queued. Using
  tegra_dma_is_empty might not be correct as one DMA chanel can be serviced
  by multiple clients.

With this change, i can browse the BT sometimes. It takes some retires to
enable the BT.

Also during the operation of the BT, i still see the messages like
"out-of-order packet arrived"

Don't know if this is an issue with ther UART or the BT protocol.

Change-Id: I1502ad8178c3c95b4128c564c9f3c1d957195d46