tegra: Hooked pm_power_off callback to the Board adaptation.
When the systemn is powered off, pm_power_off callback is called which
sets the voltage rail corresponding to the GUID NV_VDD_SoC_ODM_ID to Off.

So, to use the power off feature, implement the NV_VDD_SoC_ODM_ID GUID to power
rail mapping and PMU odm adaptation to turn off that rail.

Change-Id: I4c439325cd14bb848f100b3e7ddc029a20cd0c9d