tegra: Use write-combined DMA buffer for Tx DMA instead of zero-copy DMA.
Given the constraints of Tx DMA to only use the dword aligned buffer start
and dword aligned size, it would prudent to allocate a uncached DMA buffer
and copy the data to that buffer before the trasmit.

Main reason if Flushing L2 is not a cheap operation and the UART driver
model is having to do all the operations while holding the spin-lock
will not work.

Change-Id: I0ca14d5809258e2a62111f0a8a5e8a8053de1a34