nvmap: change from platform device to mem device
in order to implement user vs super-user protections for the nvmap
functionality, multiple device nodes need to be created with separate
file operations, similar to the mem vs kmem separation.

since nvmap is conceptually similar to the other "mem" class devices,
the multiple device nodes are added as new minor numbers in the mem
class, and the existing platform driver registration is removed.

Change-Id: Ie260e4c56679682133d90a69c0985eaaa5a4d071