UBUNTU: [Config] generic-pae switch to M586TSC
BugLink: http://bugs.launchpad.net/bugs/519448

Xen dom-U support is currently not available in the 32bit kernels.
This is because the generic-pae kernel is only an M586 kernel which
does not support the TSC, a required element for XEN support.  From the
available documentation it seems that all cpus which support PAE must
also have support for TSC.  This means that we can switch the generic-pae
kernel from M586 to M586TSC safely.  In M586TSC most XEN may then be
enabled successfully.  Enable M586TSC and thereby XEN.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>