UBUNTU: SAUCE: aufs: Fix header files inclusion in debug.h
BugLink: http://bugs.launchpad.net/bugs/517151

The aufs module fails to build with CONFIG_AUFS_DEBUG=y
The issue is not present in lucid because the whole driver code was synced
with the upstream aufs tree.

As I discussed with Andy Whitcroft on IRC, I'm not sure it justifies an SRU,
as it is unlikely that anyone would really want to build aufs with debugging
support from the karmic kernel sources package.

Signed-off-by: Mirsal Ennaime <mirsal.ennaime@gmail.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>