tegra usbh: whistler suspend/resume failure fix
With OTG disabled on whistler the suspend/resume code was hanging. The
reason is because register reads hang when clocks are disabled. This fix
tries to skip the suspend calls when device is already in suspend state.
This avoids attempts to read register with clocks disabled.

Bug 651253 : [Whistler] : USB gadget and/or EHCI bus suspend/resume
API causes crash 1st after suspend/resume

Reviewed by: vmoganty

Tested on : whistler with suspend/resume framework enabled. Suspend/resume
sequence seems to work fine on whistler and harmony with OTG
disabled in kernel and odm_query.

Change-Id: Iedc096648deab5c3b80350446c4ee67f625df27e