tegra nvddk: fixing null phy host issue when connecting to an OTG device
When the null phy host connects to an OTG device, the ID pin must set to 1 which
indicates the AP is the host. Otherwise, the device may act as host and AP will
not able to recognize the device. Also configure the pinmux according to the odmdata.

Bug 625086: [AP20 \ Whistler \ Android] ULPI Null phy doesn't work with linux
ehci driver

Change-Id: I5f1c53cbcf91a8978eab8be63431794df6e57e3a