tegra pm: implement suspend_ops
We implement suspend_ops->valid, suspend_ops->begin,
suspend_ops->enter and suspend_ops->end.

For suspend_ops->begin :
Check if CPU1 is in reset or not.
If not, reset it and maintain this state.

For suspend_ops->end :
Check if CPU1 was reseted by suspend_ops->begin.
If yes, take it out of reset.
For harmony board, add a timed wake lock as a WAR.

For suspend_ops->enter :
Do a odm query for lowest power state.
Call the appropriate low power state (LP1/LP0)
routine based on SOC type.

Change-Id: I6f4be75978573bba94e95723c023abc091e06d1e