tegra suspend: fix gpio interrupt wakeups for LP1
the top-level interrupt for the GPIO controller (i.e., at the GIC) must
be enabled in order to detect any enabled GPIO interrupts in the LP1
suspend state.

this change still has a problem that if a GPIO interrupt is only used
as a wakeup source, and is never configured through the ODM kit, the
system is unlikely to go into suspend at all because the interrupt
will never be cleared (so after it fires once, it will continue firing
every time the system attempts to suspend).

enable SD card insert / removal as a wakeup source on harmony to test
GPIO wakeups

Change-Id: I66293cbd3fcf0667ee5c965d002c8722f0fc0811