tegra uart: Removing the modification of the mctrl
The serial_core have the modem state variable mctrl which was getting
modified by the tegra uart driver. The low level driver should not modify
this variable as serial core uses this for the modem control statesi and uses
in future for controling the modem states.

Removing the modification of this variable from low level driver.

Tested with suspend/resume with BT on whistler by monitoring the modem signals.

Change-Id: Ib6244c5160e40623a40e4764f5f156dc4f386ec9