tegra pm: enable wake from EC keyboard (for harmony)
Register the gpio to which EC wake up is connected as input
and register it's interrupt with RM.

Device wakes up properly but display does not turn on. This is
probably because, the above gpio only wakes up AP but does not
send the key-code packet to AP. Subsequent key press sends the
key-code and hence turns on the display.

Tested on harmony + r04 firmware of EC.
Reviewed by bnihlani and nmalwade.

Change-Id: I1befd95f0799098f2ada12772b5321c190fffe0e