tegra mmc: Fixed data getting copied on write protected sdcard
The issue was due to incorrect card present check in sdhci-tegra.c
because of which the ro status of card was always returned as -1. This
implied that host does not support reading ro switch and the card was
assumed to be write enabled.

Fixed by correcting the card present check.

Bug 657078 ([AP20/Android/Whistler/Storage]: Able to copy data even
though sdcard is write-protected)