tegra uart: Make device id equal to instance id
Changed the registration to use the device ID same as instance.
With this change instance 0 will show up as ttyHS0 and instance 1
will show up as ttyHS1 and so on. Before this change, if the instance 0
was not used on a platform, instance 1 would have showed up as ttyHS0.

Bug 656451
Tested with Harmony.
Made sure that all nodes showup in the device list except the missing instance.
Change-Id: Ib4e04b12f16002deb899b38630de102c24e588b0
Reviewed-on: http://git-master/r/735
Reviewed-by: Ramachandrudu Kandhala <rkandhala@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>