tegra usb: Device hangs after connecting devices to USB3
AP20 pop chip is not having USB3 ID pin. Currently both pop and
non-pop whistler are using same odm query with Idpin type
as "Cable Id". Since, it is not reflecting correct Id pin type for
pop, causing the issue.

With this change based on the board type, proper id pin type is
returned as part of usb odm query for USB3. Without ID pin dynamic
detection of devices is not possible, so host functionality is
disabled for USB3 on pop boards. With this change USB3 host is
working for non-pop modules and no hang is seen with pop modules.

Bug 655520 : [AP20/Android/Whistler]: Device hangs after connecting
             USB devices(mouse/keyboard).

Tested on : whistler E1109 and  whistlerE1008

Change-Id: I5cd67c54756666e58cbc778da2b1015df27c14c6
Reviewed-on: http://git-master/r/795
Reviewed-by: Narendra Damahe <ndamahe@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>