tegra: Tristating the unsed pin groups in VDDIO_NAND rail.
The power numbers were high on the VDDIO_NAND rail, when the Nand module
was not present in EMMC boot. Doing a Pullup of the unused pin groups
and tristating the pad groups in the VDDIO_NAND rail, brings down the
power from 9mW to 2mW.

Bug 630271

Tested Nand and EMMC boot up on whistler.

Change-Id: I98ac3c0ae237093b0f17552799e0beb956c735e9
Reviewed-on: http://git-master/r/822
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>