nvmap: add barriers around atomic_read and atomic_set
add SMP memory barriers before atomic_read and after atomic_set to
ensure that any changes to the values are properly flushed so that
they are visible to all processors in the system, since neither
operator guarantees atomicity.

change the poison flag to be a non-zero value, so that zero-filled
memory isn't improperly detected as a valid handle.

re-enable error prints for pinning & unpinning handles with no local
references, now that the performance regression-causing misuse in the
codec firmware is resolved.

Change-Id: Ibbf5f0ad48f83f5eb475afb6dec8633adba8e3ca
Reviewed-on: http://git-master/r/844
Reviewed-by: Antti Hatala <ahatala@nvidia.com>
Tested-by: Antti Hatala <ahatala@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>