tegra power: simplify lp2, enable for non-wakelock builds
since the target lp2 time is being computed directly from the OS
timer list, it can be used computed immediately prior to entering
lp2.  previously, it was computed by a previous invocation of the
idle loop and stored in a static variable; which was clearly wrong
since, after the calculation, the idle loop would wfi.

move all of the lp2 logic out of the CONFIG_WAKELOCK block; the
only information wakelocks provide is an additional hint as to
whether or not entering lp2 is allowed, so this can be checked
at the top of the idle loop and non-wakelock dependent code can
handle the actual entry into LP2.

partial fix for bug 660382

Change-Id: I397c8dbc946cb3495e9b9b247d11eae782503623
Reviewed-on: http://git-master/r/846
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Tested-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: Trivikram Kasivajhula <tkasivajhula@nvidia.com>
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>