tegra power: implement early_suspend handler for display power
For proper power behavior Android expects the display to be
turned off and on in the early_suspend stages instead of
as a part of the actual suspend sequence.

Add notifications for display on/off in the nvrm power
notification interface, and call them at appropriate times.
When CONFIG_HAS_EARLYSUSPEND is not defined, continue to
notify nvrm_daemon in the power manager suspend/resume calls.

Tabify & whitespace cleanup nvrm_user.c.

Change-Id: Ie0129e0a5812352e97c95c4f323928d907782dee
Reviewed-on: http://git-master/r/866
Reviewed-by: Gary King <gking@nvidia.com>
Tested-by: Gary King <gking@nvidia.com>